+++ packages/plugins/eslint-plugin-react-hooks-extra/src/rules/ensure-use-memo-has-non-empty-deps.spec.ts
@@ -126,7 +126,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -147,7 +147,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -168,7 +168,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -189,7 +189,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -210,7 +210,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -231,7 +231,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -252,7 +252,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -273,7 +273,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -294,7 +294,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "roact",
         },
       },
@@ -315,7 +315,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -336,7 +336,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -357,7 +357,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           importSource: "@pika/react",
         },
       },
@@ -378,7 +378,7 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
       settings: {
-        reactOptions: {
+        "react-x": {
           additionalHooks: {
             useMemo: ["useCustomMemo"],
           },